Generalize worker threads.
yrmcds::worker is generalized and renamed cybozu::worker.
It now accepts any kind of jobs as callback functions.
